// src/shared/ui/Button/consts/enums.ts
export enum ButtonTheme {
  CLEAR = "clear",
  OUTLINE = "outline",
  OUTLINE_RED = "outline_red",
  BACKGROUND = "background",
  BACKGROUND_INVERTED = "backgroundInverted",
  CLEAR_INVERTED = "clear_inverted",
}

export enum ButtonSize {
  M = "size_m",
  S = "size_s",
  XL = "size_xl",
  L = "size_l",
}